Conditional density estimation with HAL in a single cross-validation fold

Conditional density estimation with HAL in a single cross-validation fold

cv_haldensify(
  fold,
  long_data,
  wts = rep(1, nrow(long_data)),
  lambda_seq = exp(seq(-1, -13, length = 100))
)

Arguments

fold
Object specifying cross-validation folds as generated by a call to make_folds.

long_data
A data.table or data.frame object containing the data in long format, as given in Díaz I, van der Laan MJ (2011). “Super learner based conditional density estimation with application to marginal structural models.” The International Journal of Biostatistics, 7(1), 1--20. , as produced by format_long_hazards.

wts
A numeric vector of observation-level weights, matching in its length the number of records present in the long format data. Default is to weight all observations equally.

lambda_seq
A numeric sequence of values of the tuning parameter of the Lasso L1 regression passed to fit_hal.