Confidence Intervals for IPW Estimates of the Causal Effects of Stochatic Shift Interventions

Confidence Intervals for IPW Estimates of the Causal Effects of Stochatic Shift Interventions

# S3 method for ipw_haldensify
confint(object, parm = seq_len(object$psi), level = 0.95, ...)

Arguments

object
An object of class ipw_haldensify, produced by invoking the function ipw_shift, for which a confidence interval is to be computed.

parm
A numeric vector indicating indices of object$est for which to return confidence intervals.

level
A numeric indicating the nominal level of the confidence interval to be computed.

...
Other arguments. Not currently used.

Value

A named numeric vector containing the parameter estimate from a ipw_haldensify object, alongside lower/upper Wald-style confidence intervals at a specified coverage level.

Details

Compute confidence intervals for estimates produced by ipw_shift.

Examples

# simulate data
n_obs <- 50
W1 <- rbinom(n_obs, 1, 0.6)
W2 <- rbinom(n_obs, 1, 0.2)
W3 <- rpois(n_obs, 3)
A <- rnorm(n_obs, (2 * W1 - W2 - W1 * W2), 2)
Y <- rbinom(n_obs, 1, plogis(3 * A + W1 + W2 - 2 * W3 - W1 * W3))

# fit the IPW estimator
est_ipw_shift <- ipw_shift(
  W = cbind(W1, W2, W3), A = A, Y = Y, delta = 0.5,
  lambda_seq = exp(seq(-1, -10, length = 500L)),
  # arguments passed to hal9001::fit_hal()
  max_degree = 3,
  smoothness_orders = 0,
  num_knots = NULL,
  reduce_basis = 1 / sqrt(n_obs),
  # ...continue arguments for IPW
  undersmooth_type = "dcar"
)

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

#> `smoothness_orders` set externally but overridden to 0.

confint(est_ipw_shift)

#> # A tibble: 2 × 3
#>   lwr_ci   est upr_ci
#>    <dbl> <dbl>  <dbl>
#> 1  0.158 0.369  0.644
#> 2  0.157 0.378  0.665