Print Method for Highly Adaptive Lasso Conditional Density Estimation

Print Method for Highly Adaptive Lasso Conditional Density Estimation

# S3 method for haldensify
print(x, ...)

Arguments

x
An object of class haldensify.

...
Other options (not currently used).

Value

None. Called for the side effect of printing an informative summary of slots of objects of class haldensify.

Details

The print method for objects of class haldensify

Examples

# simulate data: W ~ U[-4, 4] and A|W ~ N(mu = W, sd = 0.5)
set.seed(429153)
n_train <- 50
w <- runif(n_train, -4, 4)
a <- rnorm(n_train, w, 0.5)
# learn relationship A|W using HAL-based density estimation procedure
haldensify_fit <- haldensify(
  A = a, W = w, n_bins = c(3, 5),
  lambda_seq = exp(seq(-1, -10, length = 50)),
  max_degree = 3, smoothness_orders = 0, reduce_basis = 0.1
)
print(haldensify_fit)

#> HAL Conditional Density Estimation

#>   

#> Number of bins over support of A: 5

#>   

#> CV-selected lambda: 0.0037

#>   

#> Summary of fitted HAL:

#> Warning: Coefficients for many lamdba exist --
#> Summarizing coefficients corresponding to minimum lambda.

#>           coef               term
#>  1:   5.679759        (Intercept)
#>  2:  12.420725 [ I(bin_id >= 1) ]
#>  3: -12.080531 [ I(W >= -3.182) ]
#>  4:  11.667281 [ I(bin_id >= 2) ]
#>  5: -10.939450 [ I(W >= -1.172) ]
#>  6: -10.401103  [ I(W >= 0.866) ]
#>  7: -10.391974 [ I(W >= -1.864) ]
#>  8:  10.240475 [ I(bin_id >= 3) ]
#>  9:  10.111291 [ I(W >= -3.663) ]
#> 10:  -9.907410 [ I(W >= -3.853) ]